Fix a panic due to out-of-bounds access of Events
Due to forgetting to take into account the additional offset from the event pane height, it was possible
to overflow the number of events in the list when attempting to display one.
This was exacerbated by an issue when PgDown'ing near the bottom of the pane, which erroneously added
too much to the self.events_position.